Extract KeepAlive logic into its own class
